Change position of about button to navigation
